testModels.py to pytest and re-include in CI (#1032)
